server/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.115.12
h11==0.14.0
idna==3.10
numpy==2.2.5
opencv-python==4.11.0.86
packaging==25.0
pdf2image==1.17.0
pillow==11.2.1
pydantic==2.11.3
pydantic_core==2.33.1
pytesseract==0.3.13
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
